Fix TSK07/18/20: skins/pigs should not be blocked there, because it will stop production of the pigfarm
the problem is in the blocked pig skins in the storage, that is why after some time pigfarms will have 5 skins there (serf can't bring them to the storage, because they are blocked). So pigfarm will not produce anything in that situation.
This is an AI issue (it can't produce food or leather because of that and thus AI can't work in the mission as it suppose to - he does not produce enough leather soldiers of just dies of hunger, I don't recall what was the exact issue)